+++ Plugins/RawCMS.Plugins.Core/Model/FieldClientValidation.cs
@@ -1,0 +1,8 @@
+﻿namespace RawCMS.Plugins.Core.Model
+{
+    public class FieldClientValidation
+    {
+        public string Name { get; set; }
+        public string Function { get; set; }
+    }
+}
