+++ Plugins/RawCMS.Plugins.Core/Model/FieldInfo.cs
@@ -1,0 +1,11 @@
+﻿using RawCMS.Library.Schema;
+using System.Collections.Generic;
+
+namespace RawCMS.Plugins.Core.Model
+{
+    public class FieldInfo
+    {
+        public FieldType Type { get; set; }
+        public List<FieldClientValidation> Validations { get; set; }
+    }
+}
