+++ RawCMS.Library/Schema/Validation/BaseJavascriptValidator.cs
@@ -7,35 +7,6 @@ using System.Collections.Generic;
 
 namespace RawCMS.Library.Schema.Validation
 {
-    public class NullSafeDict<TKey, TValue> : Dictionary<TKey, TValue> where TValue : class
-    {
-        public new TValue this[TKey key]
-        {
-            get
-            {
-                if (!ContainsKey(key))
-                {
-                    return null;
-                }
-                else
-                {
-                    return base[key];
-                }
-            }
-            set
-            {
-                if (!ContainsKey(key))
-                {
-                    Add(key, value);
-                }
-                else
-                {
-                    base[key] = value;
-                }
-            }
-        }
-    }
-
     public abstract class BaseJavascriptValidator : FieldTypeValidator
     {
         public abstract string Javascript { get; }
